create vr-user component
